LayoutTests/platform/blackberry/fast/dom/Window/window-resize-and-move-arguments-expected.txt

This test makes sure that we don't allow calling the window moving and resizing methods with less than 2 arguments and that we do allow calling them with more.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



window.resizeTo Tests

Testing - resizeTo with 0 arguments
PASS window.outerWidth is resetWidth
PASS window.outerHeight is resetHeight
Testing - resizeTo with 1 argument
PASS window.outerWidth is resetWidth
PASS window.outerHeight is resetHeight
Testing - resizeTo with more than 2 arguments
FAIL window.outerWidth should be 100. Was 1024.
FAIL window.outerHeight should be 200. Was 600.

window.resizeBy Tests

Testing - resizeBy with 0 arguments
PASS window.outerWidth is resetWidth
PASS window.outerHeight is resetHeight
Testing - resizeBy with 1 argument
PASS window.outerWidth is resetWidth
PASS window.outerHeight is resetHeight
Testing - resizeBy with more than 2 arguments
FAIL window.outerWidth should be 1049. Was 1024.
FAIL window.outerHeight should be 650. Was 600.

window.moveTo Tests

Testing - moveTo with 0 arguments
PASS window.screenX is resetX
PASS window.screenY is resetY
Testing - moveTo with 1 argument
PASS window.screenX is resetX
PASS window.screenY is resetY
Testing - moveTo with more than 2 arguments
FAIL window.screenX should be 25. Was 0.
FAIL window.screenY should be 50. Was 0.

window.moveBy Tests

Testing - moveBy with 0 arguments
PASS window.screenX is resetX
PASS window.screenY is resetY
Testing - moveBy with 1 argument
PASS window.screenX is resetX
PASS window.screenY is resetY
Testing - moveBy with more than 2 arguments
FAIL window.screenX should be 25. Was 0.
FAIL window.screenY should be 50. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
